requirements.txt

streamlit==1.36.0
yfinance==0.2.65
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
matplotlib==3.8.4
requests==2.31.0
google-generativeai==0.8.5